Add tests for clearAfterFailure and clearAfterSuccess

+++ test/src/main/kotlin/net/twisterrob/gradle/test/GradleRunnerRule.kt
@@ -72,7 +72,7 @@ open class GradleRunnerRule : TestRule {
 					success = true
 				} finally {
 					tearDown()
-					if ((success && needClearAfterSuccess) || needClearAfterFailure) {
+					if ((success && needClearAfterSuccess) || (!success && needClearAfterFailure)) {
 						temp.delete()
 					}
 				}
